requirements.txt
#pip install -r requirements.txt

django
pillow
django-admin-interface
django-crispy-forms
django-passwords
djangorestframework
django-progressive-web-app
mysqlclient
pymysql
requests
